fix: write-chain-info after u17
