+++ packages/builders/scripts/orchestration/write-chain-info.js
@@ -1,0 +1,13 @@
+import { makeHelpers } from '@agoric/deploy-script-support';
+
+/** @type {import('@agoric/deploy-script-support/src/externalTypes.js').CoreEvalBuilder} */
+export const defaultProposalBuilder = async () =>
+  harden({
+    sourceSpec: '@agoric/orchestration/src/proposals/init-chain-info.js',
+    getManifestCall: ['getManifestForChainInfo'],
+  });
+
+export default async (homeP, endowments) => {
+  const { writeCoreEval } = await makeHelpers(homeP, endowments);
+  await writeCoreEval('gov-orchestration', defaultProposalBuilder);
+};
