+++ packages/vm-config/decentral-itest-orchestration-config.json
@@ -8,6 +8,7 @@
         "@agoric/builders/scripts/vats/init-localchain.js",
         "@agoric/builders/scripts/vats/init-transfer.js",
         "@agoric/builders/scripts/vats/init-orchestration.js",
+        "@agoric/builders/scripts/orchestration/write-chain-info.js",
         {
             "module": "@agoric/builders/scripts/inter-protocol/init-core.js",
             "entrypoint": "defaultProposalBuilder",
